winsys/amdgpu: allocate IBs like normal buffers
There is a big IB buffer whose size is 256KB and normal IBs are allocated
from it. Each driver command stream (CS) has its own big IB buffer.

The maximum size of allocated IBs from the big buffer is 64KB.
(so there's a minimum of 4 IBs per buffer)

However, the size is determined when flushing, so if the used size is only
1KB, the next IB will start after the 1KB.

After the big IB buffer is all used, another one is allocated or reused.
The reusing is done by the buffer allocator itself.

Reviewed-by: Christian König <christian.koenig@amd.com>